Add newline transcoders.
For future reference: MRI appears to generate newline.c, the
transcoder logic and data, into the root of the Ruby source tree.
As a result, our scripts did not find it. I moved the .c file and
ran the scripts, but future updates will need to this or something
like it.